Squash bugs, fix windows not running well, added examples
Former-commit-id: 5b3500abd54a77e889fd727d9a67e74a03127836

+++ legoify.py
@@ -1,12 +1,13 @@
-from subprocess import call
+from PIL import GifImagePlugin
 from PIL import Image
+from subprocess import call
 import shutil
 import sys
 import os
 
 brickh = 30
 brickw = 30
-brick = "brick30.png"
+brick = "brick.png"
 
 cmdargs = sys.argv
 
@@ -75,10 +76,19 @@ def makeLegoImage( baseImage ):
 			legoImage.paste( makeLegoBrick( bp[0], bp[1], bp[2] ), ( x * brickw, y * brickh, ( x + 1 ) * brickw, ( y + 1 ) * brickh ) )
 	return legoImage
 
-# open gif to start splitting
-gifImage = Image.open( filename )
+# check if image is animated
+def is_animated(im):
+	try:
+		im.seek(1)
+		return True
+	except EOFError:
+		return False
 
-newSize = gifImage.size
+
+# open gif to start splitting
+baseImage = Image.open( filename )
+newSize = baseImage.size
+static = filename.lower().endswith( ".gif" ) and is_animated( baseImage )
 
 # scale image
 scale = 1
@@ -91,28 +101,39 @@ if newSize[0] > 30 or newSize[1] > 30:
 	
 	newSize = ( int( round( newSize[0] / scale ) ), int( round( newSize[1] / scale ) ) )
 
-# check if dir exists, if not, make it
-if not os.path.exists( "./tmp_frames/" ):
-	os.makedirs( "./tmp_frames/" )
+if( static ):
+	print( "Animated gif detected, will now legofy each frame and recreate the gif and save as lego_{}".format(filename) )
+	# check if dir exists, if not, make it
+	if not os.path.exists( "./tmp_frames/" ):
+		os.makedirs( "./tmp_frames/" )
 
-# for each frame in the gif, save it
-for i, frame in enumerate( iter_frames( gifImage ) ):
-	frame.save( './tmp_frames/frame_{}.png'.format( ( "0" * ( 4 - len( str( i ) ) ) ) + str( i ) ), **frame.info )
+	# for each frame in the gif, save it
+	for i, frame in enumerate( iter_frames( baseImage ) ):
+		frame.save( './tmp_frames/frame_{}.png'.format( ( "0" * ( 4 - len( str( i ) ) ) ) + str( i ) ), **frame.info )
 
-# make lego images from gif
-for file in os.listdir("./tmp_frames"):
-	if file.endswith(".png"):
-		print("Working on {}".format(file))
-		im = Image.open("./tmp_frames/{}".format(file)).convert("RGBA")
-		if scale != 1:
-			im.thumbnail(newSize, Image.ANTIALIAS)
-		makeLegoImage(im).save("./tmp_frames/{}".format(file))
+	# make lego images from gif
+	for file in os.listdir( "./tmp_frames" ):
+		if file.endswith( ".png" ):
+			print( "Working on {}".format( file ) )
+			im = Image.open( "./tmp_frames/{}".format( file ) ).convert( "RGBA" )
+			if scale != 1:
+				im.thumbnail( newSize, Image.ANTIALIAS )
+			makeLegoImage( im ).save( "./tmp_frames/{}".format( file ) )
 
-# make new gif "convert -delay 10 -loop 0 *.png animation.gif"
-delay = gifImage.info["duration"] / len(os.listdir("./tmp_frames"))
-call(["convert", "-delay", str(delay/10), "-loop", "0", "./tmp_frames/*.png", "lego_" + filename])
-print("Creating gif with filename\"lego_{}\"".format(filename))
+	# make new gif "convert -delay 10 -loop 0 *.png animation.gif"
+	delay = str( baseImage.info["duration"] / 10 )
+	
+	command = "convert -delay {} -loop 0 ./tmp_frames/*.png lego_{}".format( delay, filename )
 
-shutil.rmtree('./tmp_frames')
+	print( command )
+	call( command.split( " " ) )
+	print( "Creating gif with filename\"lego_{}\"".format( filename ) )
 
-print("finished")
+	shutil.rmtree( './tmp_frames' )
+else:
+	print( "Static image detected, will now legofy and save as lego_{}".format(filename) )
+	if scale != 1:
+		im.thumbnail( newSize, Image.ANTIALIAS )
+	makeLegoImage( baseImage ).save( "lego_{}".format( filename ) )
+
+print("Finished!")
